huggingface-deploy/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
tensorflow==2.20.0
pillow==11.0.0
python-multipart==0.0.20
numpy==1.26.4
